refactor versions state code
